Adds parameter to pass home directory path on ADB command line ==============================================================
This change adds an additional AndroidDebugBridge.createBridge() method that
takes a String comprising a path to the home directory to be used by adb.
If this is not null then adb will be started with the -D option and the
path passed through.

The home directory is selected in the following order:

1) The -D option.
2) The value of the 'HOME' environment variable.
3) Built in defaults.

File modified:
   ddmlib/src/main/java/com/android/ddmlib/AndroidDebugBridge.java